fix for missing tags due to tag id inconsistency, increment version to 0.4.3
Tag ids/hashes are inconsistent in git2-rs / libgit2. Sometimes, the tag's id are separate ids of tags (find_tag), sometimes thay are the ids of the target commit.
To fix the problem, first look for the id with find_tag, then with find_commit if not found with find_tags