Feature Addition: Reverse Option (#79)
Hey mlange-42,

Just wanted to drop in a pull request for something you might find a bit
hacky, but I found myself needing a reverse option. The approach here is
a bit unconventional – instead of overhauling the code, I went for
reversing the text_lines and grid. To tackle the issue of characters
facing the wrong direction, I flipped them (so a 'left-up' becomes
'left-down', and so on).

It might be more ideal to adjust how the Character struct is accessed
with the reverse option in mind, but that would require a significant
refactor of the existing code. From what I've tested, it looks like it
works pretty well across different character sets. Would love to hear
your thoughts on this and if it's something you'd consider merging.

Cheers!