Describe algorithm for graph layout (#112)
This is a start on documentation of the main algorithm (issue #6).

I have been pushing around with the code for some time and I seem to
discover more nuances all the time. I think the full algorithm
documentation will take a significant amount of time, and therefore it
would be more sensible to write it gradually.

The goal of this PR is to describe all the main steps of the algorithm
in broad terms.